cleanup ldap sync validation
